Add files via upload
This simple website is for practice on how to create a playground environment on which we can drag html elements. On the left half of the screen are various html elements from which to choose from. You can then click and drag a particular element onto the right side of the screen into the playground (white background). To return the element to its original start position, drag the html element onto the gray side of the screen.